+++ spec/components/previews/better_ui/table/table_component_preview.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "ostruct"
+
 module BetterUi
   module Table
     class TableComponentPreview < ViewComponent::Preview